requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
blas=1.0=mkl
blosc=1.21.3=h6c2663c_0
brotli=1.0.9=h2bbff1b_7
brotli-bin=1.0.9=h2bbff1b_7
brotlipy=0.7.0=py38h2bbff1b_1003
bzip2=1.0.8=he774522_0
ca-certificates=2023.01.10=haa95532_0
certifi=2023.5.7=py38haa95532_0
cffi=1.15.1=py38h2bbff1b_3
cfitsio=3.470=h2bbff1b_7
charls=2.2.0=h6c2663c_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py38haa95532_0
cloudpickle=2.2.1=py38haa95532_0
colorama=0.4.6=py38haa95532_0
cryptography=39.0.1=py38h21b164f_0
cuda-cccl=12.1.109=0
cuda-cudart=11.7.99=0
cuda-cudart-dev=11.7.99=0
cuda-cupti=11.7.101=0
cuda-libraries=11.7.1=0
cuda-libraries-dev=11.7.1=0
cuda-nvrtc=11.7.99=0
cuda-nvrtc-dev=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-runtime=11.7.1=0
cytoolz=0.12.0=py38h2bbff1b_0
dask-core=2023.4.1=py38haa95532_0
fftw=3.3.9=h2bbff1b_1
filelock=3.9.0=py38haa95532_0
freetype=2.12.1=ha860e81_0
fsspec=2023.4.0=py38haa95532_0
giflib=5.2.1=h8cc25b3_3
icc_rt=2022.1.0=h6049295_2
idna=3.4=py38haa95532_0
imagecodecs=2021.8.26=py38h319e4f4_2
imageio=2.26.0=py38haa95532_0
importlib-metadata=6.0.0=py38haa95532_0
intel-openmp=2021.4.0=haa95532_3556
jinja2=3.1.2=py38haa95532_0
jpeg=9e=h2bbff1b_1
jxrlib=1.1=he774522_2
lazy_loader=0.1=py38haa95532_0
lcms2=2.12=h83e58a3_0
lerc=3.0=hd77b12b_0
libaec=1.0.4=h33f27b4_1
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libcublas=11.10.3.66=0
libcublas-dev=11.10.3.66=0
libcufft=10.7.2.124=0
libcufft-dev=10.7.2.124=0
libcurand=10.3.2.106=0
libcurand-dev=10.3.2.106=0
libcusolver=11.4.0.1=0
libcusolver-dev=11.4.0.1=0
libcusparse=11.7.4.91=0
libcusparse-dev=11.7.4.91=0
libdeflate=1.17=h2bbff1b_0
libffi=3.4.2=hd77b12b_6
libnpp=11.7.4.75=0
libnpp-dev=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libnvjpeg-dev=11.8.0.2=0
libpng=1.6.39=h8cc25b3_0
libtiff=4.5.0=h6c2663c_2
libuv=1.44.2=h2bbff1b_0
libwebp=1.2.4=hbc33d0d_1
libwebp-base=1.2.4=h2bbff1b_1
libzopfli=1.0.3=ha925a31_0
locket=1.0.0=py38haa95532_0
lz4-c=1.9.4=h2bbff1b_0
markupsafe=2.1.1=py38h2bbff1b_0
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py38h2bbff1b_0
mkl_fft=1.3.1=py38h277e83a_0
mkl_random=1.2.2=py38hf11a4ad_0
mpmath=1.2.1=py38haa95532_0
networkx=2.8.4=py38haa95532_1
numpy=1.24.3=py38hf95b240_0
numpy-base=1.24.3=py38h005ec55_0
opencv-python=4.7.0.72=pypi_0
openjpeg=2.4.0=h4fc8c34_0
openssl=1.1.1t=h2bbff1b_0
packaging=23.0=py38haa95532_0
partd=1.2.0=pyhd3eb1b0_1
pillow=9.4.0=py38hd77b12b_0
pip=23.1.2=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=23.0.0=py38haa95532_0
pysocks=1.7.1=py38haa95532_0
python=3.8.16=h6244533_3
pytorch=2.0.0=py3.8_cuda11.7_cudnn8_0
pytorch-cuda=11.7=h16d0643_3
pytorch-mutex=1.0=cuda
pywavelets=1.4.1=py38h2bbff1b_0
pyyaml=6.0=pypi_0
requests=2.29.0=py38haa95532_0
scikit-image=0.20.0=py38h3513d60_0
scipy=1.9.1=py38he11b74f_0
setuptools=66.0.0=py38haa95532_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.9=h6c2663c_0
sqlite=3.41.2=h2bbff1b_0
sympy=1.11.1=py38haa95532_0
tifffile=2021.7.2=pyhd3eb1b0_2
tk=8.6.12=h2bbff1b_0
toolz=0.12.0=py38haa95532_0
torchaudio=2.0.0=pypi_0
torchvision=0.13.1=cpu_py38h378ed51_0
tqdm=4.65.0=pypi_0
typing_extensions=4.5.0=py38haa95532_0
urllib3=1.26.15=py38haa95532_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.38.4=py38haa95532_0
win_inet_pton=1.1.0=py38haa95532_0
xz=5.2.10=h8cc25b3_1
yaml=0.2.5=he774522_0
zfp=0.5.5=hd77b12b_6
zipp=3.11.0=py38haa95532_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
